// frontend-angular/src/app/api/model/nFLPlayerSearchResult.ts
/**
 * Vector DB LnL demo API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Stats } from './stats';


export interface NFLPlayerSearchResult { 
    id: string;
    name: string;
    position: string;
    position_group: string;
    age?: number;
    debut_year?: string;
    headshot?: string;
    weight?: number;
    display_height?: string;
    team?: string;
    abbreviation?: string;
    stats?: Stats;
    document?: string;
    embedding?: Array<number>;
    similarity?: number;
    raw_similarity?: number;
}