// frontend-angular/src/app/api/model/nFLPlayerVectorVisualization.ts
/**
 * Vector DB LnL demo API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Coordinates } from './coordinates';


export interface NFLPlayerVectorVisualization { 
    id?: string;
    name?: string;
    position?: string;
    position_group?: string;
    team?: string;
    document?: string;
    embedding?: Array<number>;
    similarity?: number;
    raw_similarity?: number;
    coordinates?: Coordinates;
}